+++ ios/Classes/CountlyiOS/Countly.m
@@ -92,7 +92,10 @@ static dispatch_once_t onceToken;
     
     config = [self checkAndFixInternalLimitsConfig:config];
     
-  [CountlyServerConfig.sharedInstance retrieveServerConfigFromStorage:config.sdkBehaviorSettings];
+    if (config.disableSDKBehaviorSettingsUpdates) {
+        [CountlyServerConfig.sharedInstance disableSDKBehaviourSettings];
+    }
+    [CountlyServerConfig.sharedInstance retrieveServerConfigFromStorage:config.sdkBehaviorSettings];
 
     CountlyCommon.sharedInstance.maxKeyLength = config.sdkInternalLimits.getMaxKeyLength;
     CountlyCommon.sharedInstance.maxValueLength = config.sdkInternalLimits.getMaxValueSize;
@@ -167,7 +170,7 @@ static dispatch_once_t onceToken;
     CountlyDeviceInfo.sharedInstance.customMetrics = [customMetricsTruncated cly_limited:@"Custom metric"];
     
     [Countly.user save];
-  // If something added related to server config, make sure to check CountlyServerConfig.notifySdkConfigChange
+    // If something added related to server config, make sure to check CountlyServerConfig.notifySdkConfigChange
     [CountlyServerConfig.sharedInstance fetchServerConfig:config];
     
 #if (TARGET_OS_IOS)
@@ -317,6 +320,8 @@ static dispatch_once_t onceToken;
     
     if (config.indirectAttribution)
         [self recordIndirectAttribution:config.indirectAttribution];
+    
+    [CountlyHealthTracker.sharedInstance sendHealthCheck];
 }
 
 - (CountlyConfig *) checkAndFixInternalLimitsConfig:(CountlyConfig *)config
@@ -453,11 +458,13 @@ static dispatch_once_t onceToken;
 - (void)applicationWillResignActive:(NSNotification *)notification
 {
     CLY_LOG_D(@"App enters background");
+    [CountlyHealthTracker.sharedInstance saveState];
 }
 
 - (void)applicationDidEnterBackground:(NSNotification *)notification
 {
     CLY_LOG_D(@"App did enter background.");
+    [CountlyHealthTracker.sharedInstance saveState];
     [self suspend];
 }
 
@@ -469,6 +476,8 @@ static dispatch_once_t onceToken;
 - (void)applicationWillTerminate:(NSNotification *)notification
 {
     CLY_LOG_D(@"App will terminate.");
+    
+    [CountlyHealthTracker.sharedInstance saveState];
     
     CountlyConnectionManager.sharedInstance.isTerminating = YES;
     
@@ -692,12 +701,14 @@ static dispatch_once_t onceToken;
         CLY_LOG_I(@"Going out of CLYTemporaryDeviceID mode and switching back to normal mode.");
 
         [CountlyDeviceInfo.sharedInstance initializeDeviceID:deviceID];
-
+        
         [CountlyPersistency.sharedInstance replaceAllTemporaryDeviceIDsInQueueWithDeviceID:deviceID];
 
         [CountlyConnectionManager.sharedInstance proceedOnQueue];
 
         [CountlyRemoteConfigInternal.sharedInstance downloadRemoteConfigAutomatically];
+        
+        [CountlyHealthTracker.sharedInstance sendHealthCheck];
 
         return;
     }