+++ ios/Classes/CountlyiOS/CountlyConfig.h
@@ -675,6 +675,16 @@ typedef enum : NSUInteger
  */
 @property(nonatomic, copy) NSString *sdkBehaviorSettings;
 
+/**
+ * Disable the server configuration updates to the server
+ */
+@property(nonatomic) BOOL disableSDKBehaviorSettingsUpdates;
+
+/**
+ * Will disable back off mechanism
+ */
+@property (nonatomic) BOOL disableBackoffMechanism;
+
 #if (TARGET_OS_IOS)
 /**
  * Variable to access content configurations.