+++ lib/src/countly_config.dart
@@ -45,6 +45,8 @@ class CountlyConfig {
   bool _autoEnrollABOnDownload = false;
   int? _requestDropAgeHours;
   String? _sdkBehaviorSettings;
+  bool _backoffMechanismDisabled = false;
+  bool _sdkBehaviorSettingsUpdatesDisabled = false;
 
   /// instance of CountlyConfigApm
   final CountlyConfigApm _countlyConfigApmInstance = CountlyConfigApm();
@@ -137,6 +139,10 @@ class CountlyConfig {
 
   String? get sdkBehaviorSettings => _sdkBehaviorSettings;
 
+  bool get backoffMechanismDisabled => _backoffMechanismDisabled;
+  
+  bool get sdkBehaviorSettingsUpdatesDisabled => _sdkBehaviorSettingsUpdatesDisabled;
+  
   /// getter for CountlyConfigApm instance that is used to access CountlyConfigApm methods
   CountlyConfigApm get apm => _countlyConfigApmInstance;
 
@@ -380,6 +386,18 @@ class CountlyConfig {
   /// This is used to provide the server config in case the server config is not available in the initializion time
   CountlyConfig setSDKBehaviorSettings(String sdkBehaviorSettings) {
     _sdkBehaviorSettings = sdkBehaviorSettings;
+    return this;
+  }
+  
+  /// Disable the backoff mechanism
+  CountlyConfig disableBackoffMechanism() {
+    _backoffMechanismDisabled = true;
+    return this;
+  }
+
+  /// Disable the server config updates to the server
+  CountlyConfig disableSDKBehaviorSettingsUpdates() {
+    _sdkBehaviorSettingsUpdatesDisabled = true;
     return this;
   }
 }