+++ lib/src/web/plugin_config.dart
@@ -1,5 +1,5 @@
 class CountlyFlutterPluginConfig {
-  static const String SDK_VERSION_STRING = '25.4.0';
+  static const String SDK_VERSION_STRING = '25.4.1';
   static const String SDK_NAME = 'dart-flutterbnp-web';
-  static const String WEB_SDK_URL = 'https://cdn.jsdelivr.net/npm/countly-sdk-web@25.4.0/lib/countly.min.js';
+  static const String WEB_SDK_URL = 'https://cdn.jsdelivr.net/npm/countly-sdk-web@25.4.1/lib/countly.min.js';
 }