+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## 25.4.1-np
+* Added fullscreen support for feedback widgets.
+* Added "disableSDKBehaviorSettingsUpdates()" init config method to disable server config updates.
+* Improved request queue handling with a built-in backoff mechanism which is enabled by default.
+* Added "disableBackoffMechanism()" init config method to disable backoff behavior.
+* Added "attemptToSendStoredRequests()" method for events that needs quick sending.
+* Added support for SDK health checks after initialization for iOS.
+* Added timezone support for Web.
+
+* Mitigated an issue when an install referrer triggered.
+
+* Updated underlying Android SDK version to 25.4.1
+* Updated underlying iOS SDK version to 25.4.2
+* Updated underlying Web SDK version to 25.4.1
+
 ## 25.4.0-np
 * ! Minor breaking change ! Removed Secure.ANDROID_ID on Android and UIDevice.currentDevice.identifierForVendor on iOS usages in device ID generation. The SDKs now exclusively uses random UUIDs for device ID generation.
 * ! Minor breaking change ! SDK now has Server Configuration feature and it is enabled by default. Changes made on SDK Manager > SDK Configuration on your server will affect SDK behavior directly.